Make cargo-embed's rtt multicore-aware
